Sampsa dev

+++ compressai_vision/conversion/mpeg_vcm.py
@@ -160,6 +160,29 @@ def MPEGVCMToOpenImageV6(  # noqa: C901
         ...
 
 
+    ``validation_csv_file`` = ``detection_validation_labels_5k.csv`` looks like this:
+
+    ::
+
+        ImageID,LabelName,Confidence
+        0001eeaf4aed83f9,airplane,1
+        000a1249af2bc5f0,person,1
+        001083f05db4352b,car,1
+        00146ba1e50ed8d8,person,1
+        ...
+
+    --> Converted to proper OpenImageV6 format (into ``classifications.csv``):
+
+    ::
+
+        ImageID,Source,LabelName,Confidence
+        0001eeaf4aed83f9,verification,/m/0cmf2,1
+        0004886b7d043cfd,verification,/m/01g317,0
+        0004886b7d043cfd,verification,/m/04hgtk,0
+        0004886b7d043cfd,verification,/m/09j2d,0
+        ...
+
+
     ``output_directory``: Path to where the OpenImageV6 formatted files are dumped.  Files under that path are:
 
     ::
@@ -448,22 +471,27 @@ def MPEGVCMToOpenImageV6(  # noqa: C901
                     )
 
     # pathlib.Path(target_data_dir).mkdir(parents=True, exist_ok=True) # not this
+
+    if os.path.exists(target_data_dir):
+        print(
+                "WARNING: the target data_dir (image directory) already exists.  Will leave as is"
+            )
+        print("DONE!")
+        return
+    
+    if (segmentation_csv_file is not None) and os.path.exists(target_mask_dir):
+        print(
+                "WARNING: the target mask_dir (segmentation mask image directory) already exists.  Will leave as is"
+            )
+        print("DONE!")
+        return
+
     if link:
         if verbose:
             print("linking image dir", data_dir, "to", target_data_dir)
-        try:
-            os.symlink(data_dir, target_data_dir)
-        except FileExistsError:
-            print(
-                "WARNING: the target data_dir (image directory) already exists.  Will leave as is"
-            )
+        os.symlink(data_dir, target_data_dir)
         if segmentation_csv_file is not None:
-            try:
-                os.symlink(mask_dir, target_mask_dir)
-            except FileExistsError:
-                print(
-                    "WARNING: the target mask_dir (segmentation mask image directory) already exists.  Will leave as is"
-                )
+            os.symlink(mask_dir, target_mask_dir)
     else:
         if verbose:
             print(
@@ -484,3 +512,5 @@ def MPEGVCMToOpenImageV6(  # noqa: C901
                     "this might take a while..",
                 )
             shutil.copytree(mask_dir, target_mask_dir)
+
+    print("DONE!")
